+++ src/rez/utils/filesystem.py
@@ -54,7 +54,7 @@ class TempDirs(object):
             self.dirs = set()
 
         for path in dirs:
-            if os.path.exists(path):
+            if os.path.exists(path) and not os.getenv("REZ_KEEP_TMPDIRS"):
                 shutil.rmtree(path)
 
     @classmethod
@@ -82,12 +82,56 @@ def retain_cwd():
         os.chdir(cwd)
 
 
+def safe_makedirs(path):
+    # makedirs that takes into account that multiple threads may try to make
+    # the same dir at the same time
+    if not os.path.exists(path):
+        try:
+            os.makedirs(path)
+        except OSError:
+            if not os.path.exists(path):
+                raise
+
+
 def is_subdirectory(path_a, path_b):
     """Returns True if `path_a` is a subdirectory of `path_b`."""
     path_a = os.path.realpath(path_a)
     path_b = os.path.realpath(path_b)
     relative = os.path.relpath(path_a, path_b)
     return (not relative.startswith(os.pardir + os.sep))
+
+
+def copy_or_replace(src, dst):
+    '''try to copy with mode, and if it fails, try replacing
+    '''
+    try:
+        shutil.copy(src, dst)
+    except (OSError, IOError), e:
+        # It's possible that the file existed, but was owned by someone
+        # else - in that situation, shutil.copy might then fail when it
+        # tries to copy perms.
+        # However, it's possible that we have write perms to the dir -
+        # in which case, we can just delete and replace
+        import errno
+
+        if e.errno == errno.EPERM:
+            import tempfile
+            # try copying into a temporary location beside the old
+            # file - if we have perms to do that, we should have perms
+            # to then delete the old file, and move the new one into
+            # place
+            if os.path.isdir(dst):
+                dst = os.path.join(dst, os.path.basename(src))
+
+            dst_dir, dst_name = os.path.split(dst)
+            dst_temp = tempfile.mktemp(prefix=dst_name + '.', dir=dst_dir)
+            shutil.copy(src, dst_temp)
+            if not os.path.isfile(dst_temp):
+                raise RuntimeError(
+                    "shutil.copy completed successfully, but path"
+                    " '%s' still did not exist" % dst_temp)
+            os.remove(dst)
+            shutil.move(dst_temp, dst)
 
 
 def copytree(src, dst, symlinks=False, ignore=None, hardlinks=False):
